Separated passing and failing decompiler highlighting tests
